refactor(metastructure): give the drift readers their own package
Five functions that read what has moved out of band on a stack lived in the
metastructure root, where the actors that need them cannot reach them: the root
imports its subpackages, so a subpackage importing the root is a cycle. That is
what kept the rotation scheduler in the root beside its two scheduled siblings,
and it is why the root is where shared helpers accumulate.

They now live in internal/metastructure/drift, which both the apply path and any
unattended actor can import. Nothing about the reading changed; this is a move,
and the apply path's drift tests were re-run rather than assumed.

The alternative was a second copy, and there is evidence against it. The
rotation work already produced one, and deleted it: the copy had drifted from
the original by dropping an `Operation != "update"` filter. A helper reachable
from everywhere is how that happens, so there is now one reader rather than one
per caller.

What this unblocks, measured rather than claimed: `generator_rotator.go` now
references no symbol defined in the metastructure root, and neither do
`auto_reconciler.go` or `stack_expirer.go`. Extracting the three actors is
therefore mechanical and is left to its own change, since it also rewires the
supervisor and a combined diff would be harder to review than either half.

Two files in this directory were already unformatted before this branch and are
left that way: reformatting them here would put unrelated churn in a refactor.
