+++ internal/metastructure/metastructure.go
@@ -31,11 +31,11 @@ import (
 	"github.com/platform-engineering-labs/formae/internal/metastructure/changeset"
 	"github.com/platform-engineering-labs/formae/internal/metastructure/config"
 	"github.com/platform-engineering-labs/formae/internal/metastructure/discovery"
+	"github.com/platform-engineering-labs/formae/internal/metastructure/drift"
 	"github.com/platform-engineering-labs/formae/internal/metastructure/forma_command"
 	"github.com/platform-engineering-labs/formae/internal/metastructure/forma_persister"
 	"github.com/platform-engineering-labs/formae/internal/metastructure/generator_update"
 	"github.com/platform-engineering-labs/formae/internal/metastructure/messages"
-	"github.com/platform-engineering-labs/formae/internal/metastructure/patch"
 	"github.com/platform-engineering-labs/formae/internal/metastructure/policy_update"
 	"github.com/platform-engineering-labs/formae/internal/metastructure/querier"
 	"github.com/platform-engineering-labs/formae/internal/metastructure/reaping"
@@ -327,7 +327,7 @@ func (m *Metastructure) ApplyForma(forma *pkgmodel.Forma, config *config.FormaCo
 	// first, we guarantee that if no incomplete commands exist, all their resources are already
 	// persisted and visible to subsequent queries.
 	if !config.Simulate {
-		if err := m.checkForConflictingCommands(stackLabelsFromForma(forma)); err != nil {
+		if err := m.checkForConflictingCommands(drift.StackLabelsFromForma(forma)); err != nil {
 			return nil, err
 		}
 
@@ -352,12 +352,12 @@ func (m *Metastructure) ApplyForma(forma *pkgmodel.Forma, config *config.FormaCo
 	if config.Mode == pkgmodel.FormaApplyModeReconcile && config.Force {
 		assertMods := make(map[string][]datastore.ResourceModification)
 		assertWitnesses := make(map[string]json.RawMessage)
-		for _, stackLabel := range stackLabelsFromForma(forma) {
-			if err := loadModificationsAndWitnesses(m.Datastore, stackLabel, assertMods, assertWitnesses); err != nil {
+		for _, stackLabel := range drift.StackLabelsFromForma(forma) {
+			if err := drift.LoadModificationsAndWitnesses(m.Datastore, stackLabel, assertMods, assertWitnesses); err != nil {
 				return nil, err
 			}
 		}
-		forma = assertWitnessesIntoForma(forma, assertMods, assertWitnesses)
+		forma = drift.AssertWitnessesIntoForma(forma, assertMods, assertWitnesses)
 	}
 
 	fa, err := FormaCommandFromForma(forma, config, pkgmodel.CommandApply, m.Datastore, clientID, subject, subjectName, resource_update.FormaCommandSourceUser, m.Cfg.Agent.Synchronization.Interval)
@@ -388,23 +388,23 @@ func (m *Metastructure) ApplyForma(forma *pkgmodel.Forma, config *config.FormaCo
 		modificationsByStack := make(map[string][]datastore.ResourceModification)
 		witnessByKsuid := make(map[string]json.RawMessage)
 		seenStacks := map[string]bool{}
-		for _, stackLabel := range append(stackLabelsFromForma(forma), fa.GetStackLabels()...) {
+		for _, stackLabel := range append(drift.StackLabelsFromForma(forma), fa.GetStackLabels()...) {
 			if seenStacks[stackLabel] {
 				continue
 			}
 			seenStacks[stackLabel] = true
-			if err := loadModificationsAndWitnesses(m.Datastore, stackLabel, modificationsByStack, witnessByKsuid); err != nil {
+			if err := drift.LoadModificationsAndWitnesses(m.Datastore, stackLabel, modificationsByStack, witnessByKsuid); err != nil {
 				return nil, err
 			}
 		}
 		var modifiedStacks = make(map[string]apimodel.ModifiedStack)
 		for stackLabel, modifications := range modificationsByStack {
-			unabsorbed := filterUnabsorbedModifications(modifications, forma, fa)
-			unabsorbed = append(unabsorbed, witnessedMovedModifications(modifications, witnessByKsuid, forma, fa)...)
+			unabsorbed := drift.FilterUnabsorbedModifications(modifications, forma, fa)
+			unabsorbed = append(unabsorbed, drift.WitnessedMovedModifications(modifications, witnessByKsuid, forma, fa)...)
 			if len(unabsorbed) > 0 {
 				modifiedResources := make([]apimodel.ResourceModification, 0, len(unabsorbed))
 				for _, modification := range unabsorbed {
-					modifiedResources = append(modifiedResources, toAPIResourceModification(modification))
+					modifiedResources = append(modifiedResources, drift.ToAPIResourceModification(modification))
 				}
 				modifiedStacks[stackLabel] = apimodel.ModifiedStack{
 					ModifiedResources: modifiedResources,
@@ -745,39 +745,6 @@ func (m *Metastructure) ApplyForma(forma *pkgmodel.Forma, config *config.FormaCo
 	}, nil
 }
 
-// loadModificationsAndWitnesses loads one stack's drift window into the
-// submission snapshot, plus the write witness for each modified resource
-// (GetPropertiesAtLastWrite: the state formae's own last write observed).
-// Window load failures fail the submission; a witness fetch failure degrades
-// to no witness for that resource, which classifies its movement as
-// tolerated, the pre-existing behavior.
-func loadModificationsAndWitnesses(ds datastore.Datastore, stackLabel string, modificationsByStack map[string][]datastore.ResourceModification, witnessByKsuid map[string]json.RawMessage) error {
-	modifications, err := ds.GetResourceModificationsSinceLastReconcile(stackLabel)
-	if err != nil {
-		slog.Error("Failed to load modifications since last reconcile", "stack", stackLabel, "error", err)
-		return fmt.Errorf("failed to load modifications for stack %s: %w", stackLabel, err)
-	}
-	if len(modifications) == 0 {
-		return nil
-	}
-	modificationsByStack[stackLabel] = modifications
-	for _, mod := range modifications {
-		if mod.Operation != "update" || mod.Ksuid == "" {
-			continue
-		}
-		if _, done := witnessByKsuid[mod.Ksuid]; done {
-			continue
-		}
-		witness, werr := ds.GetPropertiesAtLastWrite(mod.Ksuid)
-		if werr != nil {
-			slog.Warn("Failed to load write witness for drift classification", "ksuid", mod.Ksuid, "error", werr)
-			continue
-		}
-		witnessByKsuid[mod.Ksuid] = witness
-	}
-	return nil
-}
-
 func translateToAPICommand(fa *forma_command.FormaCommand) apimodel.Command {
 	apiCommand := apimodel.Command{
 		CommandID:   fa.ID,
@@ -1034,7 +1001,7 @@ func (m *Metastructure) DestroyForma(forma *pkgmodel.Forma, config *config.Forma
 
 	// Check for conflicting commands before generating resource updates (same reasoning as ApplyForma).
 	if !config.Simulate {
-		stackLabels := stackLabelsFromForma(forma)
+		stackLabels := drift.StackLabelsFromForma(forma)
 
 		// For destroy commands, also check stacks affected by cascade deletes
 		cascadeStacks, err := m.findCascadeStackLabels(forma)
@@ -1561,7 +1528,7 @@ func (m *Metastructure) ListDrift(stack string) (*apimodel.ModifiedStack, error)
 
 	modifiedResources := make([]apimodel.ResourceModification, 0, len(modifications))
 	for _, modification := range modifications {
-		modifiedResources = append(modifiedResources, toAPIResourceModification(modification))
+		modifiedResources = append(modifiedResources, drift.ToAPIResourceModification(modification))
 	}
 
 	return &apimodel.ModifiedStack{ModifiedResources: modifiedResources}, nil
@@ -2107,19 +2074,6 @@ func (m *Metastructure) checkForReapedTargets(forma *pkgmodel.Forma) error {
 	return nil
 }
 
-// stackLabelsFromForma extracts unique stack labels from a forma's resources.
-func stackLabelsFromForma(forma *pkgmodel.Forma) []string {
-	seen := make(map[string]bool)
-	var labels []string
-	for _, r := range forma.Resources {
-		if !seen[r.Stack] {
-			seen[r.Stack] = true
-			labels = append(labels, r.Stack)
-		}
-	}
-	return labels
-}
-
 // findCascadeStackLabels returns stack labels that would be affected by cascade
 // deletes for the given forma's resources. It queries the datastore for
 // cross-stack dependents of resources being destroyed.
@@ -2317,113 +2271,6 @@ func findCascadeTargetDeletes(
 	}
 
 	return cascadeTargetUpdates, cascadeResourceUpdates, nil
-}
-
-// toAPIResourceModification converts a datastore ResourceModification into its
-// API-model counterpart. For update operations it also computes the JSON-patch
-// document describing the drift between the properties at the last reconcile
-// and the current (cloud) properties. A failed patch computation degrades to
-// label-only display — it never fails the caller.
-func toAPIResourceModification(modification datastore.ResourceModification) apimodel.ResourceModification {
-	patchDoc := json.RawMessage(nil)
-	if modification.Operation == "update" && len(modification.OldProperties) > 0 && len(modification.Properties) > 0 {
-		if p, perr := patch.DriftPatch(modification.OldProperties, modification.Properties); perr == nil {
-			patchDoc = p
-		} else {
-			slog.Warn("failed to compute drift patch", "stack", modification.Stack, "label", modification.Label, "error", perr)
-		}
-	}
-	return apimodel.ResourceModification{
-		Stack:         modification.Stack,
-		Type:          modification.Type,
-		Label:         modification.Label,
-		Operation:     modification.Operation,
-		PatchDocument: patchDoc,
-		Properties:    modification.Properties,
-		OldProperties: modification.OldProperties,
-	}
-}
-
-// filterUnabsorbedModifications returns only those modifications that have NOT been
-// absorbed into the provided forma. A modification is considered absorbed when:
-//   - The forma contains a resource with matching stack, type, and label
-//   - No resource update was generated for that resource (i.e. its properties already
-//     match the current state in the datastore)
-//
-// This prevents false drift rejection when the user has already incorporated
-// out-of-band changes into their forma (e.g. via extract) before applying.
-func filterUnabsorbedModifications(
-	modifications []datastore.ResourceModification,
-	forma *pkgmodel.Forma,
-	fa *forma_command.FormaCommand,
-) []datastore.ResourceModification {
-	// Build a set of resources that have pending updates in the FormaCommand
-	type resourceKey struct {
-		stack    string
-		typeName string
-		label    string
-	}
-	resourcesWithUpdates := make(map[resourceKey]struct{})
-	for _, ru := range fa.ResourceUpdates {
-		// A convergence-only update propagates a source movement the stored
-		// state has already absorbed (e.g. a synced secret rotation reaching
-		// its consumer). It asserts no state differing from the current one,
-		// so it must not block absorbing the modification it follows from.
-		if ru.ConvergenceOnly() {
-			continue
-		}
-		resourcesWithUpdates[resourceKey{
-			stack:    ru.StackLabel,
-			typeName: ru.DesiredState.Type,
-			label:    ru.DesiredState.Label,
-		}] = struct{}{}
-	}
-
-	// Build a set of resources present in the forma
-	formaResources := make(map[resourceKey]struct{})
-	// A forma resource that declares an `alias` covers its previous
-	// label too. Index aliases by (stack, type, alias) so a drift recorded
-	// under the old label is absorbed when the forma renames the resource.
-	formaAliases := make(map[resourceKey]struct{})
-	for _, r := range forma.Resources {
-		formaResources[resourceKey{
-			stack:    r.Stack,
-			typeName: r.Type,
-			label:    r.Label,
-		}] = struct{}{}
-		if r.Alias != "" {
-			formaAliases[resourceKey{
-				stack:    r.Stack,
-				typeName: r.Type,
-				label:    r.Alias,
-			}] = struct{}{}
-		}
-	}
-
-	var unabsorbed []datastore.ResourceModification
-	for _, mod := range modifications {
-		key := resourceKey{
-			stack:    mod.Stack,
-			typeName: mod.Type,
-			label:    mod.Label,
-		}
-		// A modification is absorbed if:
-		// 1. The resource is present in the forma, AND
-		// 2. No resource update was generated for it (properties match current state)
-		_, inForma := formaResources[key]
-		_, hasUpdate := resourcesWithUpdates[key]
-		if inForma && !hasUpdate {
-			continue // absorbed
-		}
-		// Alias-aware absorption. A modification keyed by the OLD
-		// label is absorbed by a forma resource declaring `alias = <old>`.
-		// The rename update (if any) takes the modification with it.
-		if _, isAlias := formaAliases[key]; isAlias {
-			continue
-		}
-		unabsorbed = append(unabsorbed, mod)
-	}
-	return unabsorbed
 }
 
 func formaTouchesStacks(forma *forma_command.FormaCommand, stackLabels []string) bool {
